KMM-8204 :: Feature :: MVI implementation

+++ sample-core/src/commonMain/kotlin/com/mobilejazz/kmmsample/core/feature/hackerposts/data/mapper/ListIntToHackerNewsPostsIdsMapper.kt
@@ -3,6 +3,6 @@ package com.mobilejazz.kmmsample.core.feature.hackerposts.data.mapper
 import com.harmony.kotlin.data.mapper.Mapper
 import com.mobilejazz.kmmsample.core.feature.hackerposts.domain.model.HackerNewsPostsIds
 
-internal class ListIntToHackerNewsPostsIdsMapper : Mapper<List<Int>, HackerNewsPostsIds> {
-  override fun map(from: List<Int>): HackerNewsPostsIds = HackerNewsPostsIds(from.toList())
+internal class ListIntToHackerNewsPostsIdsMapper : Mapper<List<Long>, HackerNewsPostsIds> {
+  override fun map(from: List<Long>): HackerNewsPostsIds = HackerNewsPostsIds(from.toList())
 }
