+++ sample-core/src/commonMain/kotlin/com/mobilejazz/kmmsample/core/feature/hackerposts/domain/interactor/GetHackerNewsPostInteractor.kt
@@ -5,12 +5,13 @@ import com.harmony.kotlin.domain.interactor.either.GetInteractor
 import com.harmony.kotlin.error.HarmonyException
 import com.mobilejazz.kmmsample.core.feature.hackerposts.domain.HackerNewsQuery
 import com.mobilejazz.kmmsample.core.feature.hackerposts.domain.model.HackerNewsPost
+import kotlinx.coroutines.delay
 
 class GetHackerNewsPostInteractor(
   private val getHackerNewsPostInteractor: GetInteractor<HackerNewsPost>
 ) {
 
-  suspend operator fun invoke(hackerNewsPostId: Int): Either<HarmonyException, HackerNewsPost> {
+  suspend operator fun invoke(hackerNewsPostId: Long): Either<HarmonyException, HackerNewsPost> {
     return getHackerNewsPostInteractor(HackerNewsQuery.GetPost(hackerNewsPostId))
   }
 }
